es/requirements.txt
pandas==2.2.3
tqdm==4.66.5
elasticsearch==7.17.0
python-dateutil==2.8.2
